Squashed commit of the following:
commit 2bb84e4d4eb120ce02626198d490fec268d52708
Merge: 9520e2a 1a455c9
Author: Blackhand14 <103857356+Blackhand14@users.noreply.github.com>
Date:   Thu Oct 20 21:30:11 2022 -0500

    Merge branch 'crud_sneakers' into develop

commit 9520e2a46f744d3752349e13583fd279dc1e6792
Author: Angel Diaz <76696590+anngeldiaz24@users.noreply.github.com>
Date:   Thu Oct 20 20:40:10 2022 -0500

    Crud Empleados completo (Back y Front) (#1)

    * crud_empleados |create|index|show|stores

    * Updated

    * crud_empleado finished

    * crud_empleados finished

    * Crud completo Empleado (Back y Front)

commit 1a455c91451a01d05aa022b0ecc089394b6f29b4
Author: Blackhand14 <fercho550.medina@gmail.com>
Date:   Tue Oct 4 15:15:01 2022 -0500

    Update y Delete

commit 43303af211f5ed2c530ecc1e80d65d5be3170006
Author: Blackhand14 <fercho550.medina@gmail.com>
Date:   Thu Sep 29 15:36:00 2022 -0500

    CRD (SHOW, CREATE, INDEX)

+++ database/migrations/2022_09_26_213438_create_sneakers_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('nombre'); 
             $table->string('marca');
             $table->integer('precio');
-            $table->float('talla');
+            $table->decimal('talla');
             $table->integer('stock');
         });
     }
